third_party/blink/web_tests/platform/win/external/wpt/url/url-origin.any.worker-expected.txt

This is a testharness.js-based test.
Found 10 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Origin parsing: <a:	 foo.com> against <http://example.org/foo/bar>
  assert_equals: origin expected "null" but got "file://"
[FAIL] Origin parsing: <\\\\x\\hello> against <http://example.org/foo/bar>
  assert_equals: origin expected "http://x" but got "file://"
[FAIL] Origin parsing: <c:/foo> against <http://example.org/foo/bar>
  assert_equals: origin expected "null" but got "file://"
[FAIL] Origin parsing: <h://.> without base
  assert_equals: origin expected "null" but got "file://"
[FAIL] Origin parsing: <http://!"$&'()*+,-.;=_`{}~/> without base
  assert_equals: origin expected "http://!\\"$&'()*+,-.;=_`{}~" but got "http://!\\"$&'()%2A+,-.;=_`{}~"
[FAIL] Origin parsing: <blob:ftp://host/path> without base
  assert_equals: origin expected "null" but got "ftp://host"
[FAIL] Origin parsing: <blob:ws://example.org/> without base
  assert_equals: origin expected "null" but got "ws://example.org"
[FAIL] Origin parsing: <blob:wss://example.org/> without base
  assert_equals: origin expected "null" but got "wss://example.org"
[FAIL] Origin parsing: <wss://!"$&'()*+,-.;=_`{}~/> without base
  assert_equals: origin expected "wss://!\\"$&'()*+,-.;=_`{}~" but got "wss://!\\"$&'()%2A+,-.;=_`{}~"
[FAIL] Origin parsing: <w://x:0> without base
  assert_equals: origin expected "null" but got "file://"
Harness: the test ran to completion.
